requirements.txt
streamlit
yfinance
requests
transformers
kiteconnect
numpy
pandas
xgboost
scikit-learn
statsmodels
matplotlib
seaborn
vaderSentiment
python-dotenv
tensorflow
tf-keras
openai==0.28
joblib
apscheduler
pytz